backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.12.1
redis==5.0.1
slowapi==0.1.9
pandas==2.1.3
openpyxl==3.1.2
email-validator==2.1.0
python-magic==0.4.27
gunicorn==21.2.0
